requirement.txt
asgiref==3.2.3
Django==3.0.2
djangorestframework==3.11.0
get==2019.4.13
post==2019.4.13
public==2019.4.13
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
sqlparse==0.3.0
